requirements.txt
numpy
opencv-python-headless
pillow
matplotlib
scikit-learn
openai
pydantic

# python==3.9.2